+++ src/test/php/peer/BsdSocketTest.php
@@ -30,6 +30,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @test
+     * @deprecated  since 4.0.0, will be removed with 5.0.0
      */
     public function hasGivenDomain()
     {
@@ -129,7 +130,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
     public function typeDefaultsToSOCK_STREAM()
     {
         $socket = createBsdSocket(SocketDomain::$AF_UNIX, '/tmp/mysocket');
-        $this->assertEquals(SOCK_STREAM, $socket->getType());
+        $this->assertEquals(SOCK_STREAM, $socket->type());
     }
 
     /**
@@ -138,13 +139,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
     public function typeCanBeSetToSOCK_DGRAM()
     {
         $socket = createBsdSocket(SocketDomain::$AF_UNIX, '/tmp/mysocket');
-        $this->assertEquals(SOCK_DGRAM, $socket->setType(SOCK_DGRAM)->getType());
-        $this->assertSame($socket, $socket->setType(SOCK_RAW));
-        $this->assertEquals(SOCK_RAW, $socket->getType());
-        $this->assertSame($socket, $socket->setType(SOCK_SEQPACKET));
-        $this->assertEquals(SOCK_SEQPACKET, $socket->getType());
-        $this->assertSame($socket, $socket->setType(SOCK_RDM));
-        $this->assertEquals(SOCK_RDM, $socket->getType());
+        $this->assertEquals(SOCK_DGRAM, $socket->setType(SOCK_DGRAM)->type());
     }
 
     /**
@@ -153,7 +148,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
     public function typeCanBeSetToSOCK_RAW()
     {
         $socket = createBsdSocket(SocketDomain::$AF_UNIX, '/tmp/mysocket');
-        $this->assertEquals(SOCK_RAW, $socket->setType(SOCK_RAW)->getType());
+        $this->assertEquals(SOCK_RAW, $socket->setType(SOCK_RAW)->type());
     }
 
     /**
@@ -162,7 +157,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
     public function typeCanBeSetToSOCK_SEQPACKET()
     {
         $socket = createBsdSocket(SocketDomain::$AF_UNIX, '/tmp/mysocket');
-        $this->assertEquals(SOCK_SEQPACKET, $socket->setType(SOCK_SEQPACKET)->getType());
+        $this->assertEquals(SOCK_SEQPACKET, $socket->setType(SOCK_SEQPACKET)->type());
     }
 
     /**
@@ -171,7 +166,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
     public function typeCanBeSetToSOCK_RDM()
     {
         $socket = createBsdSocket(SocketDomain::$AF_UNIX, '/tmp/mysocket');
-        $this->assertEquals(SOCK_RDM, $socket->setType(SOCK_RDM)->getType());
+        $this->assertEquals(SOCK_RDM, $socket->setType(SOCK_RDM)->type());
     }
 
     /**
@@ -258,7 +253,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
     {
         $socket = createBsdSocket(SocketDomain::$AF_UNIX, '/tmp/mysocket');
         $this->assertEquals('default',
-                            $socket->getOption('bar', 'baz', 'default')
+                            $socket->option('bar', 'baz', 'default')
         );
     }
 
@@ -270,7 +265,7 @@ class BsdSocketTest extends \PHPUnit_Framework_TestCase
         $socket = createBsdSocket(SocketDomain::$AF_UNIX, '/tmp/mysocket');
         $this->assertEquals('foo',
                             $socket->setOption('bar', 'baz', 'foo')
-                                   ->getOption('bar', 'baz', 'default')
+                                   ->option('bar', 'baz', 'default')
         );
     }